게시물 수정 코드 액션으로 위치 이동

+++ app/actions/postActions.ts
@@ -38,8 +38,69 @@ export async function createPost(formData: FormData) {
   redirect("/list");
 }
 
+export async function updatePost(id: string, formData: FormData) {
+  const title = formData.get("title");
+  const content = formData.get("content");
+
+  if (!title || !content) {
+    throw new Error("입력 필요");
+  }
+
+  await connectToDb();
+
+  const cookieStore = await cookies();
+
+  const token = cookieStore.get("token")?.value;
+
+  if (!token) {
+    throw new Error("로그인 필요");
+  }
+
+  const user = jwt.verify(token, process.env.JWT_SECRET!) as {
+    userId: string;
+    nickname: string;
+  };
+
+  const post = await Post.findById(id);
+
+  if (!post) {
+    throw new Error("게시글 없음");
+  }
+
+  if (post.authorId !== user.userId) {
+    throw new Error("수정 권한 없음");
+  }
+
+  await Post.findByIdAndUpdate(id, { title, content });
+
+  revalidatePath(`/detail/${id}`);
+  redirect(`/detail/${id}`);
+}
+
 export async function deletePost(id: string) {
   await connectToDb();
+
+  const cookieStore = await cookies();
+  const token = cookieStore.get("token")?.value;
+
+  if (!token) {
+    throw new Error("로그인 필요");
+  }
+
+  const user = jwt.verify(token, process.env.JWT_SECRET!) as {
+    userId: string;
+    nickname: string;
+  };
+
+  const post = await Post.findById(id);
+
+  if (!post) {
+    throw new Error("게시글 없음");
+  }
+
+  if (post.authorId !== user.userId) {
+    throw new Error("삭제 권한 없음");
+  }
 
   await Post.findByIdAndDelete(id);
